pyproject.toml
[project]
name = "v1-finance-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.1.1",
    "prettytable>=3.17.0",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "tavily-python>=0.7.23",
]
